m216240 - Update site links
This commit updated the home README links so that they all point
somewhere. It updates the README files for hw, labs, and units
so that they accurately represent the current directories. The
home README links now point to the directories instead of the .md
files, because this will allow students to see the contents of the
directory, but still see the README files.